Deal with dead genes with duplicate names in export.
 - Split the transacting of the "current_status" tsv file into two
   steps:
    1. Transact all non-dead gene data
    2. Transact all dead gene data (one-by-one)

Step 2. checks for the existence of names, removing any that are
already allocated.  This is done "one by one" since some dead
genes share names with other dead genes (at time of writing).

Note that this is likely an error in GeneACe/Export data, and
the "one-by-one" nature of this second step will likely be removed in
a future iteration.

The splitting of the current_status transactions into two steps is
necessary, since in GeneACe, name are not removed from dead genes once
killed.
Thus, often (but not always),  a live gene and dead gene exist within the ACeDB database, both
containing a reference to the same name.