Reworked gene summary provenance changes.
Broken by an earlier refactoring.

Added a new attribute `:t` for each item in the `:history` collection
returned in the summary; this is wall clock time of the transaction (txInstant).

Use the Datomic Log API in conjunction with the `tx-data` datalog fn
to efficiently retrieve change attributes for a given transaction.

Updated tests that use the query-provenance API to use the new
signature.

Renamed the registered name of the spec used to describe the gene
summary response.

Fixes #185