Refactoring: use "short" names for request and responses parameters.
The service, in the request parameters consumed by it,
and in the response payloads returned from it, now do
not contain the clojure-esque namespace qualifier.
This is to allow for generic entities, since specs cannot be
statically defined.

A generic "entity" spec is defined in wormbase.specs.entity,
and all requests and response payloads for entities (other than for
gene) will use this.

At the backend, these short names are qualfiied (inferred) by the
endpoint for which they are associated,
such that entities in the datomic database remain the same (e.g :variation/id)