Reversed the side of the reference gene merges are stored into.
- Previously assigned to the :gene/merges on the "from" side.
- Now assign to :gene/merges on the into side.

The into gene can result in a null transaction when no changes are
propagated (no data associated with transaction, provenance only) when
the gene merge is cloned -> cloned with no properties changed.

The from gene will always result in transaction since the status is
changed (from live dead).

Record :gene/merges on the "into" gene such that a change always
yields a data transaction for a gene merge.