Secured (API) token verifications against DB
Only tokens which have been stored identically as auth-token in the DB are now considered valid.
All other tokens provided are verified against google's API and must be valid (and not expired yet) to allow access.
When no token is stored in the DB yet, the verified and valid token will be stored in the DB,
effectively becoming a permanently valid token which can be used for (scripted) API access.