Removed unmodified response from stats endpoint.
Every authorization now causes a transaction. So DB is always modified, making the "unmodified" return functionality and test void
(as every call must be authorized, therefor the DB is modified with every API call).
In order to restore this functionality one would need to: update etag generation to be more representative for state of summary objects, not entire DB.
  => Get most recent transaction for each entity type reported and generate etag from collection of those transactions.